+++ app/scripts/_old/inventory.js
