+++ app/scripts/_old/logo.js
