+++ app/scripts/items/item.js
@@ -14,7 +14,7 @@ export default class Item extends Phaser.GameObjects.Image {
     this.setDisplaySize(this.width*ratio, this.height*ratio);
     this.setOrigin(0);
     this.name = name;
-    this.setName(name)
+    this.setName(name);
     this.description = description;
     this.type = type;
     this.bolted = bolted;
@@ -22,10 +22,10 @@ export default class Item extends Phaser.GameObjects.Image {
 
 
     this.setInteractive({ useHandCursor: false  });
-    const player = scene.registry.get('player')
+    const player = scene.registry.get('player');
 
     this.on('pointerdown', () => {
-      let mode = scene.registry.get('mode')
+      let mode = scene.registry.get('mode');
       switch(mode) {
       case 'take':
         player.take(this);
