+++ app/scripts/scenes/wrapper.js
@@ -19,7 +19,7 @@ export default class Wrapper extends Phaser.Scene {
   create(/* data */) {
 
     this.player = this.add.existing(new Player(this, 'base'));
-    this.registry.set('player', this.player)
+    this.registry.set('player', this.player);
     const clothes = this.add.existing(new Item({
       scene: this,
       x: 0,
@@ -49,11 +49,8 @@ export default class Wrapper extends Phaser.Scene {
     this.scene.launch('Workshop');
     this.scene.launch('Basement');
 
-    const basement = this.scene.get('Basement');
-    const workshop = this.scene.get('Workshop');
 
     this.registry.set('room', 'Basement');
-    //this.scene.setVisible(false, workshop);
     this.scene.sleep('Workshop');
   }
